Changed the filname extension wrapping character to be [ instead of (
This is to avoid issues when the user downloads the same file twice and it renames it on their file system with a (1) appended. When this happened, and the user uploaded this latest file to the platform, the app would think the file extension is 1 instead of bimpk because it expects the file extension to be wrapped in (). By wrapping the extension with [] instead, we avoid this issue.
[] is generally rarer in typical file naming so it's safer.