Fixed stats on map and side panel in Portfolio page
*PortfolioDetails.jsx: *
1. the openEcs property added to the sites when computing the chart data seemed top be wrongly formatted.
1.a. It was looking into `properties` but `item` did not have it.
1.b. Parenthesis is mandatory to avoid skipping the addition of APPROVED count.
1.c. I removed the array wrapping the openECs as it was not explicitely comparing the wrapped value

*mmv_config.mjs:*
We are now showing the EC count by status of the Site in the site tooltip

*omapi_entities.mjs*:
Initially we were getting all the EC logs for a single ECid at a unit and rolling up the counts per status at the site level. This meant that multiple Closed statuses would count only for one.
I changed that so the Site count per status is the sum of the Site's Units counts per status.